trolling

+++ cogs/echo.py
@@ -5,7 +5,8 @@ smarkbot_staff = (390956680705474571,  # toxic
                   293079974787678209,  # smark
                   226646760515305473,  # witt
                   288028423031357441,  # zniks
-                  265085929809510402)  # maggo
+                  265085929809510402,  # maggo
+                  294936820595163142)  # bigos
 
 
 class echo(commands.Cog):
@@ -15,6 +16,8 @@ class echo(commands.Cog):
     @commands.command()
     async def echo(self, ctx, *, message):
         if ctx.message.author.id in smarkbot_staff:
+            if '@everyone' in message: 
+                return await ctx.send('Ty kurwa dziwko jebana co chcesz everyone pingować, wypierdalaj')
             await ctx.message.delete()
             await ctx.send(message)
